Extend flags package field from smallint to integer, we need more bits now

+++ sql.d/create_schema.sql
@@ -379,7 +379,7 @@ CREATE TABLE packages (
 	version text NOT NULL,
 	versionclass smallint,
 
-	flags smallint NOT NULL,
+	flags integer NOT NULL,
 	shadow bool NOT NULL,
 
 	flavors text[]
